requirements.txt
# Core ML and Deep Learning
tensorflow-cpu>=2.16.0
scikit-learn>=1.3.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1

# Data Science
numpy>=1.26.0
pandas>=2.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Web Application
streamlit>=1.29.0

# Additional Dependencies
scipy>=1.11.0
numba>=0.59.0
